Filters updated and tests added
